refactor: move package to shared runx
