+++ internal/engine/headless/pf_test.go
@@ -12,7 +12,7 @@ import (
 	rtrun "github.com/unkn0wn-root/resterm/internal/engine/runtime"
 	"github.com/unkn0wn-root/resterm/internal/httpclient"
 	"github.com/unkn0wn-root/resterm/internal/restfile"
-	"github.com/unkn0wn-root/resterm/internal/runfail"
+	"github.com/unkn0wn-root/resterm/internal/runx/fail"
 )
 
 func TestExecuteProfilePreservesWarmupStatsAndFailures(t *testing.T) {
