+++ internal/runx/report/json.go
