+++ internal/runx/report/text.go
