+++ internal/runx/report/text_profile.go
