+++ internal/runx/report/text_test.go
