+++ internal/runx/view/view_test.go
